swagger specification update v03
