+++ doonut-app-external-api/src/main/kotlin/com/doonutmate/example/configuration/ExampleConfig.kt
@@ -1,0 +1,18 @@
+package com.doonutmate.example.configuration
+
+import org.springdoc.core.models.GroupedOpenApi
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class ExampleConfig {
+    @Bean
+    fun getExampleApi(): GroupedOpenApi {
+        return GroupedOpenApi.builder()
+            .group("example")
+            .pathsToMatch("/example/**")
+            .pathsToExclude("")
+            .packagesToScan("com.doonutmate.example.controller")
+            .build()
+    }
+}
